+++ HISTORY.rst
@@ -30,6 +30,7 @@ History
     * Change briefy.common.db.mixins.identifiable.GUID.id column to be binary (default and less alocation space) (rudaporto).
     * New attributes in the local roles mixin: relationships and association proxies for easy access to the user permissions. (rudaporto).
     * New class method Base.__acl__ to map __raw_acl__ to the pyramid format (rudaporto).
+    * Improve LocalRolesMixin.add_local_role to receive a new parameter permissions to be granted (rudaporto).
  
 
 1.1.2 (2016-10-04)